Split getting and setting of terminal options
Previously, synchronizing the terminal options happened by getting the
options from the pty and setting them on the port in one if statement.
This splits these up into two separate ifs, to prepare for modifying the
settings in the next commit. This commit does not change any behaviour
(the resulting binary is likely identical).