Only use color when outputting to a terminal
Previously, log output was always colored when writing it to stdout, but
not when writing it to a file using the --log option. However, when the
output was redirected to a file using the shell, or piped to some other
process, this would introduce (probably unwanted) color escapes.

Now, color escapes are only added if the output is a terminal, as
indicated by the `isatty()` libc function.